browser-sdk: Fix missing audio when remote participant mutes
In the VideoGrid we were checking `participant.isVideoMuted`, and if so
rendering the <VideoMutedIndicator/> instead of a <GridVideoView/>

However this meant that the <audio/> element was no longer rendered, so
we pass some more props to VideoMutedIndicator and now the audio works :tada:
