+++ .changeset/modern-lies-wash.md
@@ -1,0 +1,5 @@
+---
+"@whereby.com/browser-sdk": patch
+---
+
+Fix VideoGrid missing audio when remote participant video is disabled
